+++ source/digioz.Forum/digioz.Forum/Models/ForumTopicsTrack.cs
@@ -7,7 +7,7 @@ public partial class ForumTopicsTrack
 {
     public int UserId { get; set; }
 
-    public int TopicId { get; set; }
+    public long TopicId { get; set; }
 
     public long ForumId { get; set; }
